requirements.txt
opencv-python==4.11.0.86
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.2
